Allow "use strict" at top of spec files
The globalstrict option suppresses errors that are shown if you place "use strict" outside of a function in a files that does not export anything.

The assumption is that, if you do not export anything, your file might be loaded in the global context (e.g. in a browser) but in the case of our tests that is not true.